requirements.txt
python-telegram-bot
yandex.translate